Fix compatibility issue for C/C++/C# icons
Some themes define different file types for the same things like:

   file_type_cpp <-> file_type_c++
   file_type_csharp <-> file_type_c#

Sublime Text 3 is currently not able to resolve these conflicts,
which causes the effected file types to be displayed without any icon
in the sidebar.

The corresponding core issue is filed at:

   https://github.com/SublimeTextIssues/Core/issues/2520

This commit introduces a workaround by providing icons and settings
for both definitions.