Fix installed themes search algorithm
Fixes the following issue:

If a user creates a customized overlay of an installed theme in the
User package, A File Icon creates a 'general' patch for this theme file
even if the original installed theme supports A File Icon Customization.

Example:

   Packages
     The Installed Theme Package
       .supports-a-file-icon-customization
       ThemeFile.sublime-theme

    User
       ThemeFile.sublime-theme   <- File with user customizations kills
                                    A File Icon Customization support.

Furthermore the algorithm used to find installed themes was a bit too
complicated compared to what it needs to be.

Sublime Text merges all *.sublime-theme files which have the same name.
The API `sublime.find_resources()` returns the resource files in
package loading order. We just assume the first package, which contains a theme to be the `Theme Package`.