Initialize settings and icons asynchronously
To avoid ResourceNotFoundError after calling 'Package Control: Enable Package' > A File Icon, the access to package related resources like
the _A File Icon.sublime-settings_ need to be delayed as it seems they
are not present at the point of time ST calls `plugin_loaded()`.

So we start moving usage of sublime.set_timeout_async() up in hierarchy with the goal to make all worker functions synchronous.