Refactor path module
The path module is the one to handle all path related information.
Therefore we move all the path related stuff from the settings module
over to the path module.

The settings module is to handle settings only!

This removes some of the weird cross references and is one step to
avoid circular imports.

As we no longer support ST2, we can safely use the `__package__`
attribute to determine the PACKAGE_NAME.