Fix icon overlay creation from sublime-package
This commit fixes an issue, which caused icons not to be extracted from
A File Icon.sublime-package. All users who installed A File Icon via
Package Control might therefore see out-dated icons only. Those who
installed recently might even not see any icon.

While enumerating through the sublime-package to find icons to extract,
all returned folders cause an empty `name` variable after splitting the path. As a result `FileNotFoundError` is raised if the icon source paths
are match the current enumeration cycle.

The solution is to extract only files which end with `.png` and continue
with the next file in case the current one was not found.