Fix black icons with Adaptive theme
The Adaptive.sublime-theme tints file icons black, if a light color
scheme is active, which causes A File Icon's multi-color icons to be
tinted as well.

This commit therefore ensures to always add the `layer0.tint` key to
the theme patches with a value of `None` if `"color": ""` is set.

This removes the tint defined by the patched theme.