Rename dummy syntax names
The real syntax "Plain Text" of a dummy should be visible in the status
bar to avoid confusion about "syntax not working".

Description

If we show "ColdFusion" as syntax in the status bar a user might think
the syntax is really available/installed and wonder why it doesn't
highlight as HTML. So we show "Plain Text (ColdFusion)".