Fix overlay cleanup
A File Icon runs on python 3.8 while PC runs on 3.3 still.

This commit makes sure to always remove the overlay path in case
A File Icon is disabled or removed.

It's a bit wonky during development as saving plugin.py causes the whole
overlay to be cleaned up, but that's an acceptable edge case.