Fix crash if no themes are present
Fixes #43

This commit fixes an issue which causes theme patching upon startup to
crash in case no themes are returned by `sublime.find_resources("*.sublime-theme")`.

It's a bit strange to happen though as the API should at least return
paths to Default.sublime-theme and Adaptive.sublime-theme which are
shipped with ST by default. The issue indicates an race condition with
API not being ready quickly enough from within the asynchronous worker
thread of the plugin_host.

To help working around that, the initial setup process is delayed by
200ms, which should hopefully help.

Furthermore theme patching is modified to no longer crash in case no
themes are found. It might still result in no icons being displayed
because no patches are created then.

This abnormal situation is therefore printed as error to console.